Fixed warning.

+++ magick/opencl.c
@@ -509,9 +509,9 @@ static char* getBinaryCLProgramName(MagickCLEnv clEnv, MagickOpenCLProgram prog,
     }
     ptr++;
   }
-  (void) FormatLocaleString(path,MaxTextExtent,"%s%s%s_%s_%02d_%08x_%d.bin",
+  (void) FormatLocaleString(path,MaxTextExtent,"%s%s%s_%s_%02d_%08x_%.20g.bin",
          GetOpenCLCachedFilesDirectory(),DirectorySeparator,prefix,deviceName,
-         (unsigned int)prog,signature,sizeof(char*)*8);
+         (unsigned int) prog,signature,(double) sizeof(char*)*8);
   name = (char*)AcquireMagickMemory(strlen(path)+1);
   CopyMagickString(name,path,strlen(path)+1);
   return name;
